+++ environment/transpiler/targets/javascript/core/eq.lisp
@@ -1,2 +1,0 @@
-(defmacro eq (a b)
-  `(%=== ,a ,b))