+++ environment/transpiler/targets/php/core/slot-value.lisp
@@ -1,5 +1,5 @@
 (fn slot-value (x n)
   (%slot-value-var x n))
 
-(fn =-slot-value (v x n)
+(fn (= slot-value) (v x n)
   (%=-slot-value-var v x n))