+++ modules/js/event/listener-methods.lisp
@@ -1,4 +1,4 @@
 (fn make-listener-methods (class-name)
   (@ [`(defmethod ,class-name ,(make-symbol (upcase _)) (fun)
-         (this.add-event-listener ,_ fun))]
+         ((%slot-value this add-event-listener) ,_ fun))]
      *all-events*))